Upgrade Celery to version 3.1.11 and redo how celery is managed
Celery was instantiated via an import where it set up all the
configuration needed to get celery running correctly. This was
only done via the import which made it very weird to do a

    from openspending.command import celery

and then magically celery worked (with an unused import). This
caused a lot of consfusion when contributors have been helping
out with code cleanup and remove that unused import (because an
import should never be doing anything important).

Furthermore the celery tasks were all placed in the
openspending/tasks/__init__.py file which is not the first place
one looks for important tasks.

This commit upgrades celery to version 3.1.11 and swaps the code
over to use the celery app to handle tasks and run the celery
workers. The celery app is created in openspending/tasks/celery.py
and the tasks have been split up into two modules in tasks:

* dataset (for dataset tasks)
* generic (for generic tasks like the ping)

Now to run the celery worker, it is now not needed to go through
the paster command, it suffices to run

    celery -A openspending.tasks worker

For tasks that require the pylons environment to be set (database
set up etc.) the celery command now also takes a new option, -p
where the pylons configuration (.ini) file can be provided

    celery -A openspending.tasks -p development.ini worker

OpenSpending celery use is now more in line with the documentation
and recommended use by the celery documentation.